Landing Screen
After a user logs in or creates an account, they go to our default landing screen. When we create the whole site, this screen will also have a link to the login page in case they are not logged in. In this page, the user gets to input all the data they would like to tell us about where they want to travel (just a general area) and the dates as well as what weather they would like to travel in.